refactor(sandbox): reuse control cleanup evidence
